refactor(window): move 5 more methods (timer + file-open helpers) to Window
Five methods move from `impl Editor` to `impl Window`, no delegators:

- `check_diagnostic_pull_timer` (editor_accessors.rs) — fires the
  debounced `textDocument/diagnostic` re-pull against the active
  window's LSP manager. The body now reads `self.scheduled_diagnostic_pull`,
  `self.buffer_metadata`, `self.buffers`, `self.diagnostic_result_ids`,
  `self.next_lsp_request_id`, `self.lsp` directly — no more
  `self.windows.get_mut(&self.active_window).and_then(|w| ...)` ladder.

- `redirect_active_split_away_from_dock_if_needed`,
  `restore_global_file_state`, `save_file_state_on_close`
  (file_open_orchestrators.rs) — all internal helpers that only
  touched window-scoped split-manager / split-view-state /
  buffer-metadata fields. They become straight `self.X` reads on
  Window.

- `Editor::effective_active_pair` (editor_accessors.rs) was 50 lines of
  `self.windows.get(...).and_then(|w| w.splits.as_ref())` boilerplate
  duplicating the Window implementation at `Window::effective_active_pair`.
  Now it just delegates: `self.active_window().effective_active_pair()`.

Editor call sites updated:
- `editor_tick` in mod.rs
- `Editor::open_file` and `Editor::open_file_no_focus_inner`
- `Editor::close_buffer_by_id` in buffer_close.rs
