+++ crates/fresh-editor/src/app/buffer_close.rs
@@ -56,7 +56,7 @@ impl Editor {
         }
 
         // Save file state before closing (for per-file session persistence)
-        self.save_file_state_on_close(id);
+        self.active_window().save_file_state_on_close(id);
 
         // Delete recovery data for explicitly closed buffers (including unnamed)
         if let Err(e) = self.delete_buffer_recovery(id) {
